Fix integer u32

+++ src_base/xeve_eco.c
@@ -287,8 +287,8 @@ int xeve_eco_sh(XEVE_BSW * bs, XEVE_SPS * sps, XEVE_PPS * pps, XEVE_SH * sh, int
     xeve_bsw_write1(bs, sh->deblocking_filter_on);
 
     xeve_bsw_write(bs, sh->qp, 6);
-    xeve_bsw_write_se(bs, sh->qp_u_offset);
-    xeve_bsw_write_se(bs, sh->qp_v_offset);
+    xeve_bsw_write_se(bs, (u32)sh->qp_u_offset);
+    xeve_bsw_write_se(bs, (u32)sh->qp_v_offset);
 
     /* byte align */
     u32 t0 = 0;
